Add files via upload
The objective of this project is to do exploratory data analysis of IBM dataset and visualizing meaningful insights from the data set.

Also we have designed a webpage to individually test the employee attrition using ensembling technique to make a prediction whether the employee will leave the company or not.

This  study  helps  in  knowing  why  attrition  occurs,  reasons  for employee attrition, challenges faced by managers in retaining employees and also suggest some measures in retaining employees. 


The goal of this work is to analyse how objective factors inﬂuence employee attrition, in order to identify the main causes that contribute to a worker’s decision to leave a company, and to be able to predict whether a particular employee will leave the company.